new file:   Esperimento2/agent_saves1/arrays_episode_100.pkl 	new file:   Esperimento2/agent_saves1/arrays_episode_200.pkl 	new file:   Esperimento2/agent_saves1/memory_episode_100.pkl
