--- Project0/log_csv/debug_logModelNoPenaltyRegularize.csv
@@ -1,1 +1,0 @@
-Episode,Total Reward,Max Value,Loss
