--- Project5/log_csv/debug_log_sequentialCNN_NOPER.csv
@@ -1,127 +1,0 @@
-Episode,Action,Legal Moves,Q-Values,Reward,Total Reward,State,Next State,Done
-0,0,3,"[0, 1, 2]","[578.2686157226562, 336.0379943847656, 563.1449584960938, 579.2564697265625]",-1,-1,"[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]","[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]",False
-0,0,3,"[0, 1, 2]","[578.2686157226562, 336.0379943847656, 563.1449584960938, 579.2564697265625]",-1,-2,"[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]","[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]",False
-0,0,3,"[0, 1, 2]","[578.2686157226562, 336.0379943847656, 563.1449584960938, 579.2564697265625]",-1,-3,"[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]","[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]",False
-0,0,3,"[0, 1, 2]","[578.2686157226562, 336.0379943847656, 563.1449584960938, 579.2564697265625]",-1,-4,"[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]","[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]",False
-0,0,3,"[0, 1, 2]","[578.2686157226562, 336.0379943847656, 563.1449584960938, 579.2564697265625]",-1,-5,"[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]","[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]",False
-0,0,3,"[0, 1, 2]","[578.2686157226562, 336.0379943847656, 563.1449584960938, 579.2564697265625]",-1,-6,"[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]","[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]",False
-0,0,3,"[0, 1, 2]","[578.2686157226562, 336.0379943847656, 563.1449584960938, 579.2564697265625]",-1,-7,"[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]","[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]",False
-0,0,3,"[0, 1, 2]","[578.2686157226562, 336.0379943847656, 563.1449584960938, 579.2564697265625]",-1,-8,"[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]","[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]",False
-0,0,3,"[0, 1, 2]","[578.2686157226562, 336.0379943847656, 563.1449584960938, 579.2564697265625]",-1,-9,"[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]","[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]",False
-0,0,3,"[0, 1, 2]","[578.2686157226562, 336.0379943847656, 563.1449584960938, 579.2564697265625]",-1,-10,"[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]","[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]",False
-0,0,3,"[0, 1, 2]","[578.2686157226562, 336.0379943847656, 563.1449584960938, 579.2564697265625]",-1,-11,"[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]","[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]",False
-0,0,3,"[0, 1, 2]","[578.2686157226562, 336.0379943847656, 563.1449584960938, 579.2564697265625]",-1,-12,"[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]","[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]",False
-0,0,3,"[0, 1, 2]","[578.2686157226562, 336.0379943847656, 563.1449584960938, 579.2564697265625]",-1,-13,"[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]","[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]",False
-0,0,3,"[0, 1, 2]","[578.2686157226562, 336.0379943847656, 563.1449584960938, 579.2564697265625]",-1,-14,"[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]","[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]",False
-0,0,3,"[0, 1, 2]","[578.2686157226562, 336.0379943847656, 563.1449584960938, 579.2564697265625]",-1,-15,"[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]","[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]",False
-0,0,3,"[0, 1, 2]","[578.2686157226562, 336.0379943847656, 563.1449584960938, 579.2564697265625]",-1,-16,"[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]","[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]",False
-0,0,3,"[0, 1, 2]","[578.2686157226562, 336.0379943847656, 563.1449584960938, 579.2564697265625]",-1,-17,"[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]","[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]",False
-0,0,3,"[0, 1, 2]","[578.2686157226562, 336.0379943847656, 563.1449584960938, 579.2564697265625]",-1,-18,"[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]","[[0 0 0 0]
- [0 0 0 0]
- [0 2 0 0]
- [0 4 0 0]]",False
